+++ CHANGELOG.md
@@ -6,8 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 # Unreleased
 ### Changed
-- Changed `God Emperor` `Aura Of Divine Fury` from level 15 -> 14.
-- Changed `God Emperor` `Armor Saint` from level 5 -> 4.
+- Changed `Isekai Protagonist`'s `Quicked-Footed` obtained from level 15 -> 16
+- Changed `God Emperor`'s `Aura Of Divine Fury` obtained from level 15 -> 14.
+- Changed `God Emperor`'s `Armor Saint` obtained from level 5 -> 4.
 
 # [4.2.0] - 2023-03-20
 ### Added